Consolidate Python CLI implementation
